requirements.txt
python-dotenv
pymongo[ocsp]
gspread
fastapi
uvicorn
pydantic
oauth2client
pandas
xgboost